requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
groq>=0.5.0
python-dotenv>=1.0.1